Shell simulation
This program simulates a shell. The shell  prompts for a command line and running commands. Also, it  allows the redirection of standard input and standard output and supports both foreground and background processes. Additionally, the shell support three built in commands: exit, cd, and status, as well as comments, which are lines beginning with the # character.